objectforge-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
rembg==2.0.53
requests==2.32.3
<<<<<<< HEAD
python-multipart==0.0.9
=======
httpx==0.27.0
python-multipart==0.0.9
pydantic-settings==2.4.0
Pillow==10.4.0
orjson==3.10.7
opencv-python-headless==4.10.0.84
>>>>>>> origin/codex/optimize-my-page-zy1m9v
